+++ src/Symfony/Component/RateLimiter/Tests/RateLimiterBundleTest.php
@@ -21,6 +21,7 @@ use Symfony\Component\DependencyInjection\Kernel\KernelTrait;
 use Symfony\Component\DependencyInjection\Loader\Configurator\ContainerConfigurator;
 use Symfony\Component\DependencyInjection\Reference;
 use Symfony\Component\Filesystem\Filesystem;
+use Symfony\Component\HttpKernel\Controller\ArgumentResolver\MapRateLimitValueResolver;
 use Symfony\Component\Lock\LockFactory;
 use Symfony\Component\Lock\Store\InMemoryStore;
 use Symfony\Component\RateLimiter\CompoundRateLimiterFactory;
@@ -85,6 +86,11 @@ class RateLimiterBundleTest extends TestCase
         $container->register('cache.app');
         new RateLimiterBundle()->getContainerExtension()->load([[]], $container);
 
+        $definition = $container->getDefinition('argument_resolver.map_rate_limit');
+        $this->assertSame(MapRateLimitValueResolver::class, $definition->getClass());
+        $this->assertTrue($definition->hasTag('controller.argument_value_resolver'));
+        $this->assertTrue($definition->hasTag('kernel.event_subscriber'));
+
         new ContainerRemoveMissingDependenciesPass()->process($container);
         new RemoveMissingDependenciesPass()->process($container);
         $this->assertTrue($container->hasDefinition('cache.rate_limiter'));
@@ -99,6 +105,7 @@ class RateLimiterBundleTest extends TestCase
         $this->assertFalse($container->hasDefinition('limiter_builder'));
         $this->assertFalse($container->hasDefinition('cache.rate_limiter'));
         $this->assertFalse($container->hasDefinition('rate_limiter.attribute_listener'));
+        $this->assertFalse($container->hasDefinition('argument_resolver.map_rate_limit'));
     }
 }